+++ gemrb/plugins/Core/SoundMgr.h
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/SoundMgr.h,v 1.16 2004/08/29 01:19:02 divide Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/SoundMgr.h,v 1.17 2005/02/23 20:47:01 guidoj Exp $
  *
  */
 
@@ -48,14 +48,14 @@ public:
 	SoundMgr(void);
 	virtual ~SoundMgr(void);
 	virtual bool Init(void) = 0;
-	virtual unsigned long Play(const char* ResRef, int XPos = 0, int YPos = 0, unsigned long flags = GEM_SND_RELATIVE) = 0;
-	virtual unsigned long StreamFile(const char* filename) = 0;
+	virtual unsigned int Play(const char* ResRef, int XPos = 0, int YPos = 0, unsigned int flags = GEM_SND_RELATIVE) = 0;
+	virtual unsigned int StreamFile(const char* filename) = 0;
 	virtual bool Stop() = 0;
 	virtual bool Play() = 0;
 	virtual void ResetMusics() = 0;
 	virtual void UpdateViewportPos(int XPos, int YPos) = 0;
 	// update volumes (possibly on-the-fly)
-	virtual void UpdateVolume( unsigned long = GEM_SND_VOL_MUSIC | GEM_SND_VOL_AMBIENTS ) {}
+	virtual void UpdateVolume( unsigned int = GEM_SND_VOL_MUSIC | GEM_SND_VOL_AMBIENTS ) {}
 	virtual AmbientMgr *GetAmbientMgr() { return ambim; }
 protected:
 	AmbientMgr *ambim;
